Applied review comments
