+++ runtime/common/src/impls.rs
@@ -1,5 +1,6 @@
 //! Some configurable implementations as associated type for the substrate runtime.
 
+use super::constants::{CENTI_CFG, TREASURY_FEE_RATIO};
 use super::*;
 use codec::{Decode, Encode, MaxEncodedLen};
 use common_types::CurrencyId;
@@ -44,18 +45,18 @@ where
 	fn on_unbalanceds<B>(mut fees_then_tips: impl Iterator<Item = NegativeImbalance<R>>) {
 		if let Some(fees) = fees_then_tips.next() {
 			// for fees, split the destination
-			let mut split = fees.ration(
-				super::TREASURY_FEE_RATIO.deconstruct(),
-				(Perbill::one() - super::TREASURY_FEE_RATIO).deconstruct(),
+			let (treasury_amount, mut author_amount) = fees.ration(
+				TREASURY_FEE_RATIO.deconstruct(),
+				(Perbill::one() - TREASURY_FEE_RATIO).deconstruct(),
 			);
 			if let Some(tips) = fees_then_tips.next() {
 				// for tips, if any, 100% to author
-				tips.merge_into(&mut split.1);
+				tips.merge_into(&mut author_amount);
 			}
 
 			use pallet_treasury::Pallet as Treasury;
-			<Treasury<R> as OnUnbalanced<_>>::on_unbalanced(split.0);
-			<ToAuthor<R> as OnUnbalanced<_>>::on_unbalanced(split.1);
+			<Treasury<R> as OnUnbalanced<_>>::on_unbalanced(treasury_amount);
+			<ToAuthor<R> as OnUnbalanced<_>>::on_unbalanced(author_amount);
 		}
 	}
 }
@@ -76,7 +77,7 @@ impl WeightToFeePolynomial for WeightToFee {
 	type Balance = Balance;
 
 	fn polynomial() -> WeightToFeeCoefficients<Self::Balance> {
-		let p = super::CENTI_CFG;
+		let p = CENTI_CFG;
 		let q = 10 * Balance::from(ExtrinsicBaseWeight::get());
 
 		smallvec!(WeightToFeeCoefficient {
@@ -281,156 +282,34 @@ pub mod xcm {
 #[cfg(test)]
 mod tests {
 	use super::*;
-	use frame_support::{parameter_types, traits::FindAuthor, weights::DispatchClass, PalletId};
-	use frame_system::limits;
-	use sp_core::H256;
-	use sp_runtime::{
-		testing::Header,
-		traits::{BlakeTwo256, IdentityLookup},
-		Perbill,
-	};
-	use sp_std::convert::TryFrom;
+	use mock::*;
 
-	type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
-	type Block = frame_system::mocking::MockBlock<Test>;
 	const TEST_ACCOUNT: AccountId = AccountId::new([1; 32]);
-
-	frame_support::construct_runtime!(
-		pub enum Test where
-			Block = Block,
-			NodeBlock = Block,
-			UncheckedExtrinsic = UncheckedExtrinsic,
-		{
-			System: frame_system::{Pallet, Call, Config, Storage, Event<T>},
-			Authorship: pallet_authorship::{Pallet, Call, Storage, Inherent},
-			Balances: pallet_balances::{Pallet, Call, Storage, Config<T>, Event<T>},
-			Treasury: pallet_treasury::{Pallet, Call, Storage, Config, Event<T>},
-		}
-	);
-
-	parameter_types! {
-		pub const BlockHashCount: u64 = 250;
-		pub BlockWeights: limits::BlockWeights = limits::BlockWeights::builder()
-			.base_block(10)
-			.for_class(DispatchClass::all(), |weight| {
-				weight.base_extrinsic = 100;
-			})
-			.for_class(DispatchClass::non_mandatory(), |weight| {
-				weight.max_total = Some(1024);
-			})
-			.build_or_panic();
-		pub BlockLength: limits::BlockLength = limits::BlockLength::max(2 * 1024);
-		pub const AvailableBlockRatio: Perbill = Perbill::one();
-	}
-
-	impl frame_system::Config for Test {
-		type BaseCallFilter = frame_support::traits::Everything;
-		type Origin = Origin;
-		type Index = u64;
-		type BlockNumber = u64;
-		type Call = Call;
-		type Hash = H256;
-		type Hashing = BlakeTwo256;
-		type AccountId = AccountId;
-		type Lookup = IdentityLookup<Self::AccountId>;
-		type Header = Header;
-		type Event = Event;
-		type BlockHashCount = BlockHashCount;
-		type BlockLength = BlockLength;
-		type BlockWeights = BlockWeights;
-		type DbWeight = ();
-		type Version = ();
-		type PalletInfo = PalletInfo;
-		type AccountData = pallet_balances::AccountData<u64>;
-		type OnNewAccount = ();
-		type OnKilledAccount = ();
-		type SystemWeightInfo = ();
-		type SS58Prefix = ();
-		type OnSetCode = ();
-		type MaxConsumers = frame_support::traits::ConstU32<16>;
-	}
-
-	impl pallet_balances::Config for Test {
-		type Balance = u64;
-		type Event = Event;
-		type DustRemoval = ();
-		type ExistentialDeposit = ();
-		type AccountStore = System;
-		type MaxLocks = ();
-		type MaxReserves = ();
-		type ReserveIdentifier = [u8; 8];
-		type WeightInfo = ();
-	}
-
-	parameter_types! {
-		pub const TreasuryPalletId: PalletId = PalletId(*b"py/trsry");
-		pub const MaxApprovals: u32 = 100;
-	}
-
-	impl pallet_treasury::Config for Test {
-		type Currency = pallet_balances::Pallet<Test>;
-		type ApproveOrigin = frame_system::EnsureRoot<AccountId>;
-		type RejectOrigin = frame_system::EnsureRoot<AccountId>;
-		type Event = Event;
-		type OnSlash = ();
-		type ProposalBond = ();
-		type ProposalBondMinimum = ();
-		type ProposalBondMaximum = ();
-		type SpendPeriod = ();
-		type Burn = ();
-		type BurnDestination = ();
-		type PalletId = TreasuryPalletId;
-		type SpendFunds = ();
-		type MaxApprovals = MaxApprovals;
-		type WeightInfo = ();
-	}
-
-	pub struct OneAuthor;
-	impl FindAuthor<AccountId> for OneAuthor {
-		fn find_author<'a, I>(_: I) -> Option<AccountId>
-		where
-			I: 'a,
-		{
-			Some(TEST_ACCOUNT)
-		}
-	}
-	impl pallet_authorship::Config for Test {
-		type FindAuthor = OneAuthor;
-		type UncleGenerations = ();
-		type FilterUncle = ();
-		type EventHandler = ();
-	}
-
-	pub fn new_test_ext() -> sp_io::TestExternalities {
-		let mut t = frame_system::GenesisConfig::default()
-			.build_storage::<Test>()
-			.unwrap();
-		// We use default for brevity, but you can configure as desired if needed.
-		pallet_balances::GenesisConfig::<Test>::default()
-			.assimilate_storage(&mut t)
-			.unwrap();
-		t.into()
-	}
 
 	#[test]
 	fn test_fees_and_tip_split() {
-		new_test_ext().execute_with(|| {
-			let fee = Balances::issue(10);
-			let tip = Balances::issue(20);
+		TestExternalitiesBuilder::default()
+			.build()
+			.execute_with(|| {
+				const FEE: u64 = 10;
+				const TIP: u64 = 20;
 
-			assert_eq!(Balances::free_balance(Treasury::account_id()), 0);
-			assert_eq!(Balances::free_balance(TEST_ACCOUNT), 0);
+				let fee = Balances::issue(FEE);
+				let tip = Balances::issue(TIP);
 
-			DealWithFees::on_unbalanceds(vec![fee, tip].into_iter());
+				assert_eq!(Balances::free_balance(Treasury::account_id()), 0);
+				assert_eq!(Balances::free_balance(TEST_ACCOUNT), 0);
 
-			assert_eq!(
-				Balances::free_balance(Treasury::account_id()),
-				super::TREASURY_FEE_RATIO * 10
-			);
-			assert_eq!(
-				Balances::free_balance(TEST_ACCOUNT),
-				20 + (Perbill::one() - super::TREASURY_FEE_RATIO) * 10
-			);
-		});
+				DealWithFees::on_unbalanceds(vec![fee, tip].into_iter());
+
+				assert_eq!(
+					Balances::free_balance(Treasury::account_id()),
+					TREASURY_FEE_RATIO * FEE
+				);
+				assert_eq!(
+					Balances::free_balance(TEST_ACCOUNT),
+					TIP + (Perbill::one() - TREASURY_FEE_RATIO) * FEE
+				);
+			});
 	}
 }
